requirements.txt
# Project runtime dependencies
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
playwright>=1.40.0
pydantic>=1.10.0
openai>=1.30.0
python-dotenv>=1.0.1
jsonschema>=4.0.0
mcp>=1.1.0
httpx>=0.25.0
slowapi>=0.1.9

